+++ test/functional/WorkerUnityCourseTest.php
@@ -8,10 +8,11 @@ class WorkerUnityCourseTest extends UnityWebPortalTestCase
     private static string $manager_uid = "user2_org1_test";
     private static string $manager_mail = "user2@org1.test";
     private static string $courseOwnerMail = "user2+cs124@org1.test";
+    private static string $expirationDate = "1970/01/02";
 
     public function testCreateCourse()
     {
-        global $LDAP, $USER;
+        global $LDAP, $USER, $SQL;
         $this->switchUser("Blank");
         $this->assertEquals(self::$manager_uid, $USER->uid);
         $this->assertEquals(self::$manager_mail, $USER->getMail());
@@ -25,13 +26,12 @@ class WorkerUnityCourseTest extends UnityWebPortalTestCase
             self::$course_owner_name[1],
             self::$course_owner_uid,
             self::$manager_uid,
+            self::$expirationDate,
         ]);
         $stdin_file_path = getPathFromFileHandle($stdin_file);
         try {
             executeWorker("unity-course.php", stdinFilePath: $stdin_file_path);
             // error_log(implode("\n", $output_lines));
-            // our LDAP conn doesn't know about changes from subprocess
-            unset($GLOBALS["ldapconn"]);
             $this->switchUser("Admin");
             $pi_group_entry = $LDAP->getPIGroupEntry(self::$course_gid);
             $owner_user_entry = $LDAP->getUserEntry(self::$course_owner_uid);
@@ -46,10 +46,14 @@ class WorkerUnityCourseTest extends UnityWebPortalTestCase
                 [$manager->uid],
                 $pi_group_entry->getAttribute("manageruid"),
             );
+            $newExpirationDate = $SQL->getPIGroupExpirationDate(self::$course_gid);
+            $this->assertNotNull($newExpirationDate);
+            $this->assertEquals(self::$expirationDate, date("Y/m/d", $newExpirationDate));
         } finally {
             ensurePIGroupDoesNotExist(self::$course_gid);
             ensureUserDoesNotExist(self::$course_owner_uid);
             unlink($stdin_file_path);
+            $SQL->removePIGroupExpirationDate(self::$course_gid);
         }
     }
 }