+++ tools/docker-dev/sql/bootstrap.sql
@@ -23,3 +23,8 @@ CREATE TABLE `requests` (
   `uid` varchar(128) NOT NULL,
   `timestamp` timestamp NOT NULL DEFAULT current_timestamp()
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
+
+CREATE TABLE `pi_group_expiration_dates` (
+  `gid` varchar(131) NOT NULL PRIMARY KEY,
+  `expiration_date` timestamp NOT NULL
+) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci;